+++ lib/src/sound/events/reverb.dart
@@ -1,4 +1,5 @@
 /// Provides events relating to reverb.
+import '../../game.dart';
 import '../reverb_preset.dart';
 import 'events_base.dart';
 import 'playback.dart';
@@ -6,10 +7,18 @@ import 'playback.dart';
 /// Create a reverb.
 class CreateReverb extends SoundEvent {
   /// Create a reverb.
-  const CreateReverb({required int id, required this.reverb}) : super(id);
+  const CreateReverb(
+      {required this.game, required int id, required this.reverb})
+      : super(id);
+
+  /// The game to use.
+  final Game game;
 
   /// The reverb preset to use.
   final ReverbPreset reverb;
+
+  /// Destroy this reverb.
+  void destroy() => game.sounds.add(DestroyReverb(id));
 }
 
 /// Destroy a pre-existing reverb.