+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.2.0]
+
+### Added
+
+* Added a `destroy` method to the `CreateReverb` class.
+
+### Changed
+
+* Reverbs must now be destroyed with `CreateReverb.destroy`.
+* Renamed `PlaySound.destroySound` to `destroy`.
+
+### Removed
+
+* Removed the `Game.destroyReverb` method.
+
 ## [0.1.0]
 
 ### Added